# Include a factor in add bias kernel of matmul for perf tuning